sentiment-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
transformers>=4.28.0
torch>=2.0.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
cachetools>=5.3.0
pydantic>=2.0.0 